requirements.txt
python == 3.10.6
Django == 4.1.7
djangorestframework == 3.14.0
flake8==6.0.0
python-decouple==3.7